fix_fft updated to June 22 commit
Updated to newest version to avoid confusion

+++ libs/fix_fft/src/fix_fft.cpp
@@ -103,7 +103,7 @@ inline int8_t FIX_MPY(int8_t a, int8_t b)
   
   
     /* shift right one less bit (i.e. 15-1) */
-    int c = ((int)a * (int)b) >> 6;
+    int16_t c = ((int16_t)a * (int16_t)b) >> 6;
     /* last bit shifted out = rounding-bit */
     b = c & 0x01;
     /* last shift + rounding bit */
@@ -124,9 +124,9 @@ inline int8_t FIX_MPY(int8_t a, int8_t b)
   RESULT (in-place FFT), with 0 <= n < 2**m; set inverse to
   0 for forward transform (FFT), or 1 for iFFT.
 */
-int fix_fft(int8_t fr[], int8_t fi[], int m, int inverse)
+int16_t fix_fft(int8_t fr[], int8_t fi[], int16_t m, int16_t inverse)
 {
-    int mr, nn, i, j, l, k, istep, n, scale, shift;
+    int16_t mr, nn, i, j, l, k, istep, n, scale, shift;
     int8_t qr, qi, tr, ti, wr, wi;
 
     n = 1 << m;
@@ -248,9 +248,9 @@ Serial.println("");*/
   that fix_fft "sees" consecutive real samples as alternating
   real and imaginary samples in the complex array.
 */
-int fix_fftr(int8_t f[], int m, int inverse)
+int16_t fix_fftr(int8_t f[], int16_t m, int16_t inverse)
 {
-    int i, N = 1<<(m-1), scale = 0;
+    int16_t i, N = 1<<(m-1), scale = 0;
     int8_t tt, *fr=f, *fi=&f[N];
 
     if (inverse)
